pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "easyproxy"
version = "0.1.0"
description = "一个轻量级的多协议代理服务器,支持HTTP/HTTPS/SOCKS5"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Slothtron", email = "lolioycmd@foxmail.com"}
]
keywords = ["proxy", "http", "https", "socks5", "asyncio"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: Proxy Servers",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "pydantic>=2.6.0",
    "PyYAML>=6.0.1",
    "click>=8.1.0",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/Slothtron/easyproxy-py"
Documentation = "https://github.com/Slothtron/easyproxy-py#readme"
Repository = "https://github.com/Slothtron/easyproxy-py"
Issues = "https://github.com/Slothtron/easyproxy-py/issues"

# 这是关键配置:定义命令行入口点
[project.scripts]
easyproxy = "easyproxy.cli:main"

[tool.setuptools]
packages = ["easyproxy"]

[tool.setuptools.package-data]
easyproxy = ["py.typed"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
